Terraform/Azure/apps/dataenrich_v1/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
pandas
fastparquet
pyarrow